feat: refactor adapter definitions and remove deprecated pre-alignment adapter
