+++ packages/adapters/src/define.ts
@@ -1,72 +1,59 @@
-import type { CacheableUrlRequestReturnType, MojiAdapter } from "./types";
-import semver from "semver";
+import type {
+  AdapterContext,
+  BuiltinParser,
+  v2_AdapterHandler,
+  v2_AdapterHandlerType,
+  v2_GetParseOutputFromBuiltInParser,
+  v2_OutputFn,
+} from "./types";
 
-/**
- * Validates and defines a Moji adapter configuration.
- *
- * @param {MojiAdapter} adapter - The adapter configuration object to validate
- * @throws {Error} If adapter.name is missing
- * @throws {Error} If adapter.description is missing
- * @throws {Error} If adapter.range is missing
- * @throws {Error} If adapter.range is not a valid semver range
- * @throws {Error} If required functions are missing when adapter.extend is null
- * @returns {MojiAdapter} The validated adapter configuration
- *
- * @example
- * ```ts
- * const adapter = defineMojiAdapter({
- *   name: 'my-adapter',
- *   description: 'My adapter',
- *   range: '>=1.0.0',
- * });
- * ```
- */
-export function defineMojiAdapter<
-  TMetadataUrlReturn extends CacheableUrlRequestReturnType,
-  TSequencesUrlReturn extends CacheableUrlRequestReturnType,
-  TVariationsUrlReturn extends CacheableUrlRequestReturnType,
->(adapter: MojiAdapter<
-  TMetadataUrlReturn,
-  TSequencesUrlReturn,
-  TVariationsUrlReturn
->): MojiAdapter<
-    TMetadataUrlReturn,
-    TSequencesUrlReturn,
-    TVariationsUrlReturn
+export function defineAdapterHandler<
+  TType extends v2_AdapterHandlerType,
+  TExtraContext extends Record<string, unknown>,
+  TContext extends AdapterContext,
+  TTransformOutput,
+>(
+  handler: Omit<
+    v2_AdapterHandler<TType, TExtraContext, TContext, TTransformOutput>,
+    "output"
+  > & {
+    output: v2_OutputFn<TContext, TExtraContext, TTransformOutput, any>;
+  }
+): v2_AdapterHandler<
+  TType,
+  TExtraContext,
+  TContext,
+  TTransformOutput,
+  TTransformOutput,
+  ReturnType<typeof handler.output>
+>;
+export function defineAdapterHandler<
+  TType extends v2_AdapterHandlerType,
+  TExtraContext extends Record<string, unknown>,
+  TContext extends AdapterContext,
+  TTransformOutput,
+  TAggregateOutput = TTransformOutput,
+  TOutput = TTransformOutput | TAggregateOutput,
+  TBuiltinParser extends BuiltinParser = BuiltinParser,
+  TParseOutput = v2_GetParseOutputFromBuiltInParser<BuiltinParser>,
+>(handler: v2_AdapterHandler<
+  TType,
+  TExtraContext,
+  TContext,
+  TTransformOutput,
+  TAggregateOutput,
+  TOutput,
+  TBuiltinParser,
+  TParseOutput
+>): v2_AdapterHandler<
+    TType,
+    TExtraContext,
+    TContext,
+    TTransformOutput,
+    TAggregateOutput,
+    TOutput,
+    TBuiltinParser,
+    TParseOutput
   > {
-  // validate the adapter has name, description, range.
-  if (!adapter.name) {
-    throw new Error(`adapter.name is required`);
-  }
-
-  if (!adapter.description) {
-    throw new Error(`adapter.description is required`);
-  }
-
-  if (!adapter.range) {
-    throw new Error(`adapter.range is required`);
-  }
-
-  // verify the adapter.range is a valid semver range.
-  if (semver.validRange(adapter.range) === null) {
-    throw new Error(`adapter.range is not a valid semver range ${adapter.range}`);
-  }
-
-  if (adapter.extend == null) {
-    // verify the adapter has the required functions.
-
-    // TODO: figure out how we can make it throw type error if the adapter is missing a required function
-    // const REQUIRED_FUNCTIONS = [
-    //   "sequences",
-    //   "metadata",
-    //   "variations",
-    // ] satisfies NonNullable<MojiAdapterFunctionNames<MojiAdapter>>[];
-
-    // const missingFunctions = REQUIRED_FUNCTIONS.filter((fn) => adapter[fn] == null);
-    // if (missingFunctions.length > 0) {
-    //   throw new Error(`adapter ${adapter.name} is missing required functions: ${missingFunctions.join(", ")}`);
-    // }
-  }
-
-  return adapter;
+  return handler;
 }
